fix bug in "Compare weighting methods" - exp weights
The folder "Example10- Compare Weighting Methods - new" contains the
updated results for the simulation comparing weighting methods. This
uses the new version of the exp_weights function. It also fixed a bug in
the simulation where the exp_weights were called with the argument
abs(mu) instead of mu.

The new results show that exp_weights are somewhat better than
previously, but still not as good as Bayes weights.