fix error mean not one
Under certain conditions, the mean of the Bayes weights does not equal
one. This was an issue in the bayes_weights function, due to the
incorrect normalization by q. It occurred only in the last branch of the
search for a suitable dual variable lambda.
Thanks to J. Bogdan for pointing this out.